added main functionality, base stations and sensor nodes, fixed bugs
changed all node addresses to uint16_t
changed hop_count and seq_nr to uint16_t
corrected all data to uint8_t
many more small fixes
missing: aging, good testing

+++ Ex2/MessageTypes.h
@@ -5,8 +5,8 @@
 typedef struct BroadcastMsg
 {
 	uint8_t basestation_id;
-	uint8_t seq_nr;
-	uint8_t hop_count;
+	uint16_t seq_nr;
+	uint16_t hop_count;
 	uint16_t sender_addr;
 } BroadcastMsg;
