+++ Ex2/Routingtable.h
@@ -2,10 +2,10 @@
 
 typedef struct RoutingTableEntry
 {
-	uint8_t basestation_id;
-	uint8_t mote_id;
-	uint8_t sequence_number;
-	uint8_t hop_count;
+	uint16_t basestation_id;
+	uint16_t mote_id;
+	uint16_t sequence_number;
+	uint16_t hop_count;
 	bool aging;
 	bool valid;	
 } RoutingTableEntry;
